add check for uniqueness for DepartmentService and add tests for it

+++ src/main/java/org/stepanenko/projectmanager/service/DepartmentService.java
@@ -1,11 +1,14 @@
 package org.stepanenko.projectmanager.service;
 
 import org.springframework.stereotype.Service;
+import org.stepanenko.projectmanager.exceptions.BadRequestException;
+import org.stepanenko.projectmanager.model.Client;
 import org.stepanenko.projectmanager.model.Department;
 import org.stepanenko.projectmanager.repository.DepartmentRepository;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
+import java.util.Objects;
 
 import static java.lang.String.format;
 
@@ -18,7 +21,8 @@ public class DepartmentService {
         this.departmentRepository = departmentRepository;
     }
 
-    public Department create(Department department) {
+    public Department save(Department department) {
+        verifyNameUnique(department);
         return departmentRepository.save(department);
     }
 
@@ -27,14 +31,22 @@ public class DepartmentService {
     }
 
     public Department getById(Long id) {
-        return departmentRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(format("Department by id = %s was not found")));
-    }
-
-    public Department update(Department department) {
-        return departmentRepository.save(department);
+        return departmentRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(format("Department by id = %s was not found", id)));
     }
 
     public void delete(Long id) {
+        if (!departmentRepository.existsById(id)) {
+            throw new EntityNotFoundException(format("Department with id = %s was not found", id));
+        }
         departmentRepository.deleteById(id);
+    }
+
+    private void verifyNameUnique(Department department) {
+        if (departmentRepository.findByName(department.getName())
+                .filter(e -> !Objects.equals(e.getId(), department.getId()))
+                .isPresent()) {
+            throw new BadRequestException(
+                    format("Department with name %s already exist", department.getName()));
+        }
     }
 }
